04-service | restructurization of the project
