--- backend/configs/config.yaml
@@ -1,11 +1,0 @@
-port: 5002
-log_level: debug
-log_path: /opt/dimkush_guestbook/log/
-
-database:
-    db_core: postgres
-    db_user: postgres
-    db_name: test
-    db_password: postgres
-    db_port: 5432
-    db_connections_pool: 10
