+++ configs/config.yaml
@@ -1,0 +1,17 @@
+main:
+    port: 5002
+    log_level: debug
+    log_path: /opt/dimkush_guestbook/log/
+
+audit:
+    port: 5003
+    audit_level: debug
+    log_level: error
+
+database:
+    db_core: postgres
+    db_user: postgres
+    db_name: test
+    db_password: postgres
+    db_port: 5432
+    db_connections_pool: 10
