+++ go.mod
@@ -1,0 +1,15 @@
+module github.com/DimKush/guestbook/tree/main
+
+go 1.16
+
+require (
+	github.com/gorilla/mux v1.8.0 // indirect
+	github.com/jackc/pgproto3/v2 v2.1.0 // indirect
+	github.com/lib/pq v1.6.0 // indirect
+	github.com/rs/zerolog v1.23.0 // indirect
+	golang.org/x/crypto v0.0.0-20210616213533-5ff15b29337e // indirect
+	golang.org/x/text v0.3.6 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gorm.io/driver/postgres v1.1.0 // indirect
+	gorm.io/gorm v1.21.11 // indirect
+)
